flask-backend/requirements-docker.txt
# Flask web framework
flask==3.0.0
flask-cors==4.0.0
gunicorn==21.2.0
python-dotenv==1.0.1

# Data processing
pandas==2.2.3
numpy==1.26.4
pyarrow==15.0.0

# MCAP support (ROS bag format)
mcap==1.3.0
mcap-ros2-support==0.5.6

# ML / CLIP
open-clip-torch==3.1.0
faiss-gpu==1.7.2

# PyTorch (CUDA 12.1)
--extra-index-url https://download.pytorch.org/whl/cu121
torch==2.2.0+cu121
torchvision==0.17.0+cu121

# LLM for prompt enhancement
transformers==4.47.0
safetensors==0.4.5
huggingface-hub==0.26.5

# Authentication
PyJWT==2.9.0

# Utilities
Pillow==10.2.0
tqdm==4.67.1
prompt_toolkit==3.0.51

# Testing
pytest==8.0.0
